backend/requirements1.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.7.0
moviepy>=1.0.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.8

# Auto-editor for silence removal
auto-editor>=23.0.0

# Background removal models
huggingface_hub>=0.16.0
accelerate>=0.20.0

# Subtitle generation
openai-whisper>=20230314

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Video processing
ffmpeg-python>=0.2.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
requests>=2.31.0
aiofiles>=23.1.0